refactor(sonar): harden shared state and cleanup dead code

+++ modules/common/navigation/src/main/java/fr/riege/ebsl/common/navigation/PathPlanningService.java
@@ -76,7 +76,7 @@ public final class PathPlanningService {
         int y = target.flooredY();
         int z = target.flooredZ();
         if (checker.isSolid(x, y, z)) {
-            return new PathPosition(x, (double) y + 1.0, z);
+            return new PathPosition(x, y + 1.0, z);
         }
         return target.floor();
     }
